requirements.txt
flask==2.1.3
flask-restx
werkzeug==2.1.2
python-dotenv
pymongo
bcrypt
